requirements.txt
asgiref==3.9.2
cffi==2.0.0
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-split-settings==1.3.2
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
python-dotenv==1.1.1
sqlparse==0.5.3
PyJWT==2.10.1
requests==2.32.5